refactor: DTO 폴더 경로 추가
