+++ core/src/main/kotlin/com/wespot/auth/dto/request/RefreshTokenRequest.kt
@@ -1,0 +1,5 @@
+package com.wespot.auth.dto.request
+
+data class RefreshTokenRequest(
+    val refreshToken: String
+)
