+++ core/src/main/kotlin/com/wespot/auth/dto/response/SignInResponse.kt
@@ -1,4 +1,4 @@
-package com.wespot.auth.dto
+package com.wespot.auth.dto.response
 
 data class SignInResponse (
     val accessToken: String,
