+++ core/src/main/kotlin/com/wespot/auth/dto/response/SocialResponse.kt
@@ -1,4 +1,4 @@
-package com.wespot.auth.dto
+package com.wespot.auth.dto.response
 
 data class SocialResponse(
     val socialId: String,
